feat: send log to slack
