+++ EwhaNoticeServer/settings.py
@@ -1,23 +1,8 @@
-import json
 import os
-
-from django.core.exceptions import ImproperlyConfigured
+from secret_util import get_secret
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-secret_file = os.path.join(BASE_DIR, 'secrets.json')
-with open(secret_file) as f:
-    secrets = json.loads(f.read())
-
-
-def get_secret(setting, secrets=secrets):
-    try:
-        return secrets[setting]
-    except KeyError:
-        error_msg = "Set the {} environment variable".format(setting)
-        raise ImproperlyConfigured(error_msg)
-
 
 SECRET_KEY = get_secret("SECRET_KEY")
 DEBUG = False
