+++ requirements.txt
@@ -18,6 +18,7 @@ google-cloud-storage==1.23.0
 google-resumable-media==0.5.0
 googleapis-common-protos==1.6.0
 grpcio==1.26.0
+gunicorn==20.0.4
 httplib2==0.15.0
 idna==2.8
 msgpack==0.6.2
@@ -28,6 +29,7 @@ pytz==2019.3
 requests==2.22.0
 rsa==4.0
 six==1.13.0
+slacker==0.14.0
 soupsieve==1.9.5
 sqlparse==0.3.0
 uritemplate==3.0.1
